initial mid-recording added

+++ Android/SoundRecorder/src/com/wiseapps/davacon/core/se/SERecorderStateListener.java
@@ -5,7 +5,7 @@ package com.wiseapps.davacon.core.se;
  *         Date: 4/14/14
  *         Time: 12:02 PM
  *
- * Set of public methods could not be changed!!!
+ * TODO return recordingInProgress method to its prev state (handle double position type)
  */
 public interface SERecorderStateListener {
 
@@ -17,7 +17,8 @@ public interface SERecorderStateListener {
     /**
      * Notification for update recording info
      */
-    void recordingInProgress(double position);
+//    void recordingInProgress(double position);
+    void recordingInProgress(int position);
 
     /**
      * Notification for end recording
